Move test to existing TestTorchFunctionMode class as requested by reviewer
Test validates that dynamically added functions are correctly filtered out
from get_overridable_functions() and do not appear as overridable.
